+++ src/SlmLocale/Strategy/Factory/StrategyPluginManagerFactory.php
@@ -38,48 +38,15 @@
  * @link        http://juriansluiman.nl
  */
 
-namespace SlmLocale\Service;
+namespace SlmLocale\Strategy\Factory;
 
 use Interop\Container\ContainerInterface;
-use Interop\Container\Exception\ContainerException;
-use SlmLocale\View\Helper\LocaleUrl;
-use Zend\ServiceManager\Exception\ServiceNotCreatedException;
-use Zend\ServiceManager\Exception\ServiceNotFoundException;
-use Zend\ServiceManager\FactoryInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
+use SlmLocale\Strategy\StrategyPluginManager;
 
-class LocaleUrlViewHelperFactory implements FactoryInterface
+class StrategyPluginManagerFactory
 {
-    /**
-     * Create an object
-     *
-     * @param  ContainerInterface $container
-     * @param  string $requestedName
-     * @param  null|array $options
-     *
-     * @return object
-     * @throws ServiceNotFoundException if unable to resolve the service.
-     * @throws ServiceNotCreatedException if an exception is raised when
-     *     creating a service.
-     * @throws ContainerException if any other error occurs
-     */
-    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
+    public function __invoke(ContainerInterface $container)
     {
-        $detector = $container->get('SlmLocale\Locale\Detector');
-        $request  = $container->get('Request');
-        $app      = $container->get('Application');
-
-        $match  = $app->getMvcEvent()->getRouteMatch();
-        $helper = new LocaleUrl($detector, $request, $match);
-        return $helper;
-    }
-
-    /**
-     * @param  ServiceLocatorInterface $serviceLocator
-     * @return LocaleUrl
-     */
-    public function createService(ServiceLocatorInterface $serviceLocator)
-    {
-        return $this($serviceLocator->getServiceLocator(), LocaleUrl::class);
+        return new StrategyPluginManager($container);
     }
 }